requirements.txt
appdirs==1.4.0
appnope==0.1.0
beautifulsoup4==4.5.3
decorator==4.0.11
dj-static==0.0.6
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-extensions==1.7.6
django-libsass==0.7
django-livereload-server==0.2.3
django-rosetta==0.7.12
djangorestframework==3.5.4
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
libsass==0.12.3
microsofttranslator==0.8
numpy==1.12.0
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
polib==1.0.8
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
-e git+https://github.com/0x255/pypugjs@d8fb1105551736254e0b496b592c096824dc69a7#egg=pypugjs
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
